requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
langchain
langchain-community
langchain-openai
faiss-cpu
langchain-aws
sentence-transformers
datasets
ragas
langsmith